requirements.txt
diffusers
fairscale
accelerate
tensorboard
transformers
gradio
torchdiffeq
click
torchvision
fastapi
uvicorn